dummy-ref-core-fw/teacher_console/requirements.txt
PySide6>=6.6
pyserial>=3.5
numpy>=1.24
pin>=2.7
PyYAML>=6.0
